添加了评价功能（之前没有前端页面，在订单中显示前往评价的判断好像有点问题）

+++ src/main/java/com/zhao/salechicken/controller/OrderController.java
@@ -65,7 +65,7 @@ public class OrderController {
      * @return
      */
     @DeleteMapping("/deleteOrder")
-    public R<String> deleteOrder(HttpServletRequest request, @RequestParam List<Integer> ids) {
+    public R<String> deleteOrder(HttpServletRequest request, @RequestParam List<Long> ids) {
 
         //获取当前登录用户
         Integer loginUser = BaseContext.getCurrentId();
